MME Update 172

+++ libretro-common/audio/audio_mix.c
@@ -1,4 +1,4 @@
-/* Copyright  (C) 2010-2017 The RetroArch team
+/* Copyright  (C) 2010-2018 The RetroArch team
  *
  * ---------------------------------------------------------------------------------------
  * The following license statement only applies to this file (audio_mix.c).
@@ -111,7 +111,7 @@ void audio_mix_free_chunk(audio_chunk_t *chunk)
 audio_chunk_t* audio_mix_load_wav_file(const char *path, int sample_rate)
 {
    int sample_size;
-   ssize_t len          = 0;
+   int64_t len          = 0;
    void *buf            = NULL;
    audio_chunk_t *chunk = (audio_chunk_t*)calloc(1, sizeof(*chunk));
